Skip some test if gcc is not installed

+++ testsuite/gna/bug097/testsuite.sh
@@ -2,17 +2,19 @@
 
 . ../../testenv.sh
 
-if [ -z $CC ]; then
-  CC="gcc"
+if c_compiler_is_available; then
+  if [ -z $CC ]; then
+    CC="gcc"
+  fi
+
+  $CC -c -fPIC getrand.c
+  $CC -o getrand.so --shared getrand.o
+
+  analyze tb.vhdl
+  elab_simulate tb
+
+  rm -f getrand.o getrand.so
 fi
-
-$CC -c -fPIC getrand.c
-$CC -o getrand.so --shared getrand.o
-
-analyze tb.vhdl
-elab_simulate tb
-
 clean
-rm -f getrand.o getrand.so
 
 echo "Test successful"
